// src/types/FitLogType.ts
export interface Ifitlog {
  // slice(arg0: number, arg1: number): unknown;
  id: number;
  name: string;
  image: string;
  muscleGroups: string[];
  equipment: string;
  difficulty: string;
  duration: number;
  caloriesBurned: number;
  sets: number;
  reps: string;
  rating: number;
  description: string;
  instructions: string[];
}